@@ -1,55 +1,69 @@
 {
-  "name": "navbar-dropdown-v1",
+  "name": "medi-care",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@elastic/elasticsearch": "^5.6.22",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/free-regular-svg-icons": "^5.15.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.11",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@tensorflow/tfjs-converter": "^2.7.0",
-    "@tensorflow/tfjs-core": "^2.7.0",
-    "@tensorflow/tfjs-layers": "^2.7.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "ace-css": "^1.1.0",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.1.3",
-    "d3-format": "^1.3.2",
-    "face-api.js": "^0.20.0",
-    "firebase": "^8.2.1",
-    "jquery": "^3.6.0",
-    "mapbox-gl": "^2.6.1",
-    "node-module": "^2.0.0",
-    "notistack": "^1.0.3",
-    "react": "^16.13.1",
-    "react-bootstrap": "^2.0.3",
-    "react-datepicker": "^3.3.0",
-    "react-dom": "^16.13.1",
-    "react-dropzone": "^11.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-s3-uploader": "^5.0.0",
-    "react-scripts": "^2.1.3",
-    "react-simple-chatbot": "^0.5.0",
-    "react-transition-group": "^4.4.2",
-    "restuta-react-dropzone-s3-uploader": "^0.20.4",
-    "styled-components": "^5.2.1",
-    "tslib": "^1.9.3",
-    "web-vitals": "^0.2.4"
+    "@elastic/elasticsearch": "^9.0.1",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@mui/icons-material": "^5.15.11",
+    "@mui/material": "^5.15.11",
+    "@tensorflow/tfjs-converter": "^4.22.0",
+    "@tensorflow/tfjs-core": "^4.22.0",
+    "@tensorflow/tfjs-layers": "^4.22.0",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.80",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
+    "ajv": "6.12.6",
+    "ajv-keywords": "3.5.2",
+    "axios": "^1.9.0",
+    "bootstrap": "^5.3.6",
+    "d3-format": "^3.1.0",
+    "face-api.js": "^0.22.2",
+    "firebase": "^11.6.1",
+    "formik": "^2.4.5",
+    "jquery": "^3.7.1",
+    "mapbox-gl": "^3.11.1",
+    "notistack": "^3.0.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.9",
+    "react-datepicker": "^8.3.0",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^14.3.8",
+    "react-error-boundary": "^4.0.12",
+    "react-query": "^3.39.3",
+    "react-router-dom": "^7.5.3",
+    "react-scripts": "5.0.1",
+    "react-simple-chatbot": "^0.6.1",
+    "react-transition-group": "^4.4.5",
+    "styled-components": "^4.4.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4",
+    "yup": "^1.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "eslint src/**/*.{js,jsx,ts,tsx}",
+    "format": "prettier --write src/**/*.{js,jsx,ts,tsx,css,md}",
+    "type-check": "tsc --noEmit"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
