@@ -3,7 +3,7 @@ import Navbar from './components/NavBar/Navbar';
 import './App.css';
 import Home from './components/Home/Homepage.jsx';
 import HomeHi from './components/Home/Homepagehi'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import ContactUs from './components/Health/Health';
 import Chatbot from './components/Chatbot/Chatbot';
 import DiseaesPred from './components/DiseasePred/DiseaseDetection';
@@ -19,20 +19,6 @@ import Chart from './components/chart/Chart';
 import Footer from './components/Footer/Footer.jsx';
 
 function App() {
-
-  //all the routes
-  const routes = [
-    { path: '/', Component: Home },
-    { path: '/DiseaesPred', Component: DiseaesPred },
-    { path: '/hi', Component: HomeHi },
-    { path: '/Doctor', Component: Doctor },
-    { path: '/Details', Component: Details },
-    { path: '/health', Component: ContactUs },
-    { path: '/heart-predict', Component: HeartDiseaseForm },
-    { path: '/kidney-predict', Component: KidneyDiseaseForm },
-    { path: '/report/:type', Component: DiseaseReport },
-    {path:'/covid',Component:Chart},
-  ]
 
   const [mode, setMode] = useState('light');
 
@@ -55,14 +41,12 @@ function App() {
     if (mode === 'light') {
       setMode('dark');
       document.body.style.backgroundColor = "#050A30";
-      showalert("Dark mode has been enbled", "primary");
+      showalert("Dark mode has been enabled", "primary");
     }
     else {
       setMode('light');
       document.body.style.backgroundColor = "white";
-      showalert("Light mode has been enbled", "primary");
-
-
+      showalert("Light mode has been enabled", "primary");
     }
   }
 
@@ -70,88 +54,19 @@ function App() {
     <Router>
       <Chatbot />
       <Navbar mode={mode} toggleMode={toggleMode} />
-
-      <Switch>
-
-
-        {
-          routes.map(route => {
-            const { path, Component } = route;
-
-            return <Route key={path} path={path} exact >
-              {
-                ({ match }) => (
-                  //component to make page transiiton on routing
-                  <CSSTransition
-                    in={match != null}
-                    timeout={500}
-                    unmountOnExit
-                    classNames="page"
-                  >
-                    <div className="page">
-                      <Component />
-                    </div>
-                  </CSSTransition>
-                )
-              }
-
-
-            </Route>
-          })
-        }
-
-        <Route >
-          <NotfoundPage />
-        </Route>
-
-
-        {/* <Switch>
-        <Route path='/' exact  >
-        {
-                      ({ match }) => (
-                          <CSSTransition
-                            in ={true}
-                            timeout={2000}
-                            unmountOnExit
-                            classNames="page"
-                          >
-                            <div className="page">
-                            <Home/>
-                            </div>
-                          </CSSTransition>
-                      )
-                  }
-            
-        </Route>
-        <Route path='/DiseaesPred'  > 
-                  {
-                      ({ match }) => (
-                          <CSSTransition
-                            in ={true}
-                            timeout={2000}
-                            unmountOnExit
-                            classNames="page"
-                          >
-                            <div className="page">
-                                <DiseaesPred />
-                            </div>
-                          </CSSTransition>
-                      )
-                  }
-              </Route>
-        <Route path='/hi' exact  >
-            <HomeHi/>
-        </Route>
-        <Route path='/Doctor'  > 
-            <Doctor/>
-        </Route>
-        <Route path='/Details' component={Details} />
-        <Route path='/health'  > 
-            <ContactUs/>
-        </Route>
-      </Switch> */}
-
-      </Switch>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/DiseaesPred" element={<DiseaesPred />} />
+        <Route path="/hi" element={<HomeHi />} />
+        <Route path="/Doctor" element={<Doctor />} />
+        <Route path="/Details" element={<Details />} />
+        <Route path="/health" element={<ContactUs />} />
+        <Route path="/heart-predict" element={<HeartDiseaseForm />} />
+        <Route path="/kidney-predict" element={<KidneyDiseaseForm />} />
+        <Route path="/report/:type" element={<DiseaseReport />} />
+        <Route path="/covid" element={<Chart />} />
+        <Route path="*" element={<NotfoundPage />} />
+      </Routes>
       <Footer/>
     </Router>
   );
